# Eliminate unnecessary allocations and improve performance