--- a/src/NSubstitute/Core/Arguments/IArgumentSpecification.cs
+++ b/src/NSubstitute/Core/Arguments/IArgumentSpecification.cs
@@ -7,6 +7,7 @@ namespace NSubstitute.Core.Arguments
         bool IsSatisfiedBy(object argument);
         Type ForType { get; }
         IArgumentSpecification CreateCopyMatchingAnyArgOfType(Type requiredType);
+        bool HasAction { get; }
         void RunAction(object argument);
         string DescribeNonMatch(object argument);
         string FormatArgument(object argument);