--- a/src/NSubstitute/Core/CallCollection.cs
+++ b/src/NSubstitute/Core/CallCollection.cs
@@ -8,58 +8,83 @@ namespace NSubstitute.Core
 {
     public class CallCollection : ICallCollection
     {
-        ConcurrentQueue<CallWrapper> _callWrappers = new ConcurrentQueue<CallWrapper>();
+        private ConcurrentQueue<IReceivedCallEntry> _callEntries = new ConcurrentQueue<IReceivedCallEntry>();
 
         public void Add(ICall call)
         {
             if (call == null) throw new ArgumentNullException(nameof(call));
-            _callWrappers.Enqueue(new CallWrapper(call));
+
+            IReceivedCallEntry callEntry;
+            if (call is IReceivedCallEntry callAsEntry && callAsEntry.TryTakeEntryOwnership())
+            {
+                callEntry = callAsEntry;
+            }
+            else
+            {
+                callEntry = new ReceivedCallEntry(call);
+            }
+ 
+            _callEntries.Enqueue(callEntry);
         }
 
         public void Delete(ICall call)
         {
             if (call == null) throw new ArgumentNullException(nameof(call));
 
-            var callWrapper = _callWrappers.FirstOrDefault(w => !w.IsDeleted && call.Equals(w.Call));
+            var callWrapper = _callEntries.FirstOrDefault(e => !e.IsSkipped && call.Equals(e.Call));
             if (callWrapper == null)
             {
                 throw new SubstituteInternalException("CallCollection.Delete - collection doesn't contain the call");
             }
 
-            callWrapper.Delete();
+            callWrapper.Skip();
         }
 
         public IEnumerable<ICall> AllCalls()
         {
-            return _callWrappers.ToArray()
-                .Where(w => !w.IsDeleted)
-                .Select(w => w.Call);
+            return _callEntries
+                .Where(e => !e.IsSkipped)
+                .Select(e => e.Call)
+                .ToArray();
         }
 
         public void Clear()
         {
-            //Queue doesn't have a Clear method.
-            _callWrappers = new ConcurrentQueue<CallWrapper>();
+            // Queue doesn't have a Clear method.
+            _callEntries = new ConcurrentQueue<IReceivedCallEntry>();
+        }
+ 
+        /// <summary>
+        /// Performance optimization. Allows to mark call as deleted without allocating extra wrapper.
+        /// To play safely, we track ownership, so object can be re-used only once.
+        /// </summary>
+        internal interface IReceivedCallEntry
+        {
+            ICall Call { get; }
+            bool IsSkipped { get; }
+            void Skip();
+            bool TryTakeEntryOwnership();
         }
 
         /// <summary>
         /// Wrapper to track that particular entry was deleted.
         /// That is needed because concurrent collections don't have a Delete method.
-        /// We null the hold value to not leak memory.
+        /// Notice, in most cases the original <see cref="Call"/> instance will be used as a wrapper itself.
         /// </summary>
-        private class CallWrapper
+        private class ReceivedCallEntry : IReceivedCallEntry
         {
             public ICall Call { get; private set; }
-            public bool IsDeleted => Call == null;
 
-            public CallWrapper(ICall call)
+            public bool IsSkipped => Call == null;
+
+            public void Skip() => Call = null; // Null the hold value to reclaim a bit of memory.
+
+            public bool TryTakeEntryOwnership() =>
+                throw new SubstituteInternalException("Ownership is never expected to be obtained for this entry.");
+
+            public ReceivedCallEntry(ICall call)
             {
                 Call = call;
-            }
-
-            public void Delete()
-            {
-                Call = null;
             }
         }
     }