In this PR I decided to push it to the _reasonable_ limits and decrease allocations amount during call dispatch to the bare minimum (when it makes sense, of course). Additionally, added thread-safety to the `CallActions` collection - I think that was the only remaining non-thread safe place.

Considering the following scenario:

```c#
var substitute = Substitute.For<ISomething>();
substitute.Echo(42);

// MEASURE FROM HERE
substitute.Echo(42);
```

Allocations are following:
![image](https://user-images.githubusercontent.com/1074182/56443296-b6d1e380-62fc-11e9-82ca-2d9f19267d2b.png)

The allocations under cursor come from us. I also had a plan to eliminate the enumerator, but decided to not do it, as it complicates the code (we should use our custom stack implementation instead of built-in collections) and I think it doesn't worth it. We can also do nothing about `RouteAction`, as it would be a breaking change.

The same allocation graph happens for a dispatch of already configured call.

The performance measurement is following (measured only one case when it makes much difference):

```
BEFORE

                         Method |  Job | Runtime |     Mean |    Error |   StdDev |  Gen 0 |  Gen 1 | Allocated |
------------------------------- |----- |-------- |---------:|---------:|---------:|-------:|-------:|----------:|
 DispatchInterfaceProxyCall_Int |  Clr |     Clr | 961.3 ns | 18.87 ns | 26.46 ns | 0.0629 | 0.0210 |     405 B |
 DispatchInterfaceProxyCall_Int | Core |    Core | 905.0 ns | 12.36 ns | 10.95 ns | 0.0620 | 0.0153 |     408 B |


AFTER

                         Method |  Job | Runtime |     Mean |    Error |   StdDev |  Gen 0 |  Gen 1 | Allocated |
------------------------------- |----- |-------- |---------:|---------:|---------:|-------:|-------:|----------:|
 DispatchInterfaceProxyCall_Int |  Clr |     Clr | 928.1 ns | 18.07 ns | 14.11 ns | 0.0553 | 0.0134 |     349 B |
 DispatchInterfaceProxyCall_Int | Core |    Core | 866.1 ns | 12.57 ns | 11.14 ns | 0.0505 | 0.0153 |     336 B |
```

Slight, but win without significant code complications.

@dtchepak Please review the code and let me know your opinion. If you believe that some improvements make it very complicated (like one with `ICall`) - let me know.

As usual, it's easier to review commit by commit.

Thanks!